Add files via upload
This file replaces many of the separate files that previously existed. It was used for https://github.com/phetsims/forces-and-motion-basics/issues/310 and some other issues. Having one file simplifies the artwork process so that all artwork is scaled appropriately, and all versions are in one place (original work by artist, as well as modified work for exporting).